+++ projects/netgrif-application-engine/src/lib/task-content/task-content/abstract-task-content.component.spec.ts
@@ -1,4 +1,4 @@
-import {waitForAsync, ComponentFixture, TestBed, fakeAsync, tick} from '@angular/core/testing';
+import {ComponentFixture, fakeAsync, TestBed, tick, waitForAsync} from '@angular/core/testing';
 import {MatExpansionModule} from '@angular/material/expansion';
 import {NoopAnimationsModule} from '@angular/platform-browser/animations';
 import {CommonModule} from '@angular/common';
@@ -16,7 +16,7 @@ import {LoggerService} from '../../logger/services/logger.service';
 import {TaskEventService} from '../services/task-event.service';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {DataGroupAlignment} from '../../resources/interface/data-groups';
-import {DataGroupLayoutType} from '../../resources/interface/data-group-layout';
+import {DataGroupCompact, DataGroupHideEmptyRows, DataGroupLayoutType} from '../../resources/interface/data-group-layout';
 import {IncrementingCounter} from '../../utility/incrementing-counter';
 import {BooleanField} from '../../data-fields/boolean-field/models/boolean-field';
 import {UnlimitedTaskContentService} from '../services/unlimited-task-content.service';
@@ -80,7 +80,10 @@ describe('AbstractTaskContentComponent', () => {
             expect(component.gridAreas).toBeUndefined();
 
             component.computeLayoutData([
-                createMockDataGroup([createField()], 'hello world', DataGroupAlignment.START, DataGroupLayoutType.LEGACY, true)
+                createMockDataGroup(
+                    [createField()],
+                    'hello world',
+                    DataGroupAlignment.START, DataGroupLayoutType.LEGACY, undefined, undefined, true)
             ]);
 
             expect(component.dataSource).toBeTruthy();
@@ -243,7 +246,7 @@ describe('AbstractTaskContentComponent', () => {
                         createField(true, {x: 2, y: 1, rows: 1, cols: 2}),
                         createField(true, {x: 1, y: 2, rows: 1, cols: 2}),
                         createField(true, {x: 0, y: 3, rows: 2, cols: 1})],
-                    '', DataGroupAlignment.START, DataGroupLayoutType.GRID)
+                    '', DataGroupAlignment.START, DataGroupLayoutType.GRID, DataGroupCompact.UP)
             ]);
 
             expect(component.dataSource).toBeTruthy();
@@ -259,7 +262,7 @@ describe('AbstractTaskContentComponent', () => {
             expect(grid[1][1].startsWith('xblank')).toBeTrue();
             expect(grid[1][2].startsWith('xf')).toBeTrue();
             expect(grid[1][3].startsWith('xf')).toBeTrue();
-            expect(grid[0][2] === grid[0][3]).toBeTrue();
+            expect(grid[1][2] === grid[1][3]).toBeTrue();
 
             expect(grid[2][0].startsWith('xf')).toBeTrue();
             expect(grid[3][0].startsWith('xf')).toBeTrue();
@@ -275,6 +278,172 @@ describe('AbstractTaskContentComponent', () => {
 
             expect(grid[2][3].startsWith('xblank')).toBeTrue();
             expect(grid[3][1].startsWith('xblank')).toBeTrue();
+        });
+
+        it('grid layout compact UP hide NONE', () => {
+            expect(component.dataSource).toEqual([]);
+            expect(component.gridAreas).toBeUndefined();
+
+            component.computeLayoutData([
+                createMockDataGroup([
+                        createField(true, {x: 3, y: 0, rows: 1, cols: 1}),
+                        createField(true, {x: 0, y: 2, rows: 1, cols: 2})],
+                    undefined, DataGroupAlignment.START, DataGroupLayoutType.GRID, DataGroupCompact.UP, DataGroupHideEmptyRows.NONE)
+            ]);
+
+            expect(component.dataSource).toBeTruthy();
+            expect(Array.isArray(component.dataSource)).toBeTrue();
+            expect(component.dataSource.length).toEqual(11);
+
+            expect(component.gridAreas).toBeTruthy();
+            const grid = transformStringToGrid(component.gridAreas);
+
+            expect(grid.length).toEqual(3);
+            expect(grid[0].length).toEqual(4);
+
+            expect(grid[0][0].startsWith('xf')).toBeTrue();
+            expect(grid[0][1].startsWith('xf')).toBeTrue();
+            expect(grid[0][2].startsWith('xblank')).toBeTrue();
+            expect(grid[0][3].startsWith('xf')).toBeTrue();
+            expect(grid[0][0] === grid[0][1]).toBeTrue();
+
+            for (let row = 1; row < grid.length; row++) {
+                for (let col = 0; col < grid[0].length; col++) {
+                    expect(grid[row][col].startsWith('xblank')).toBeTrue();
+                }
+            }
+        });
+
+        it('grid layout compact UP hide COMPACTED', () => {
+            expect(component.dataSource).toEqual([]);
+            expect(component.gridAreas).toBeUndefined();
+
+            component.computeLayoutData([
+                createMockDataGroup([
+                        createField(true, {x: 3, y: 0, rows: 1, cols: 1}),
+                        createField(true, {x: 0, y: 2, rows: 1, cols: 2})],
+                    undefined, DataGroupAlignment.START, DataGroupLayoutType.GRID, DataGroupCompact.UP, DataGroupHideEmptyRows.COMPACTED)
+            ]);
+
+            expect(component.dataSource).toBeTruthy();
+            expect(Array.isArray(component.dataSource)).toBeTrue();
+            expect(component.dataSource.length).toEqual(7);
+
+            expect(component.gridAreas).toBeTruthy();
+            const grid = transformStringToGrid(component.gridAreas);
+
+            expect(grid.length).toEqual(2);
+            expect(grid[0].length).toEqual(4);
+
+            expect(grid[0][0].startsWith('xf')).toBeTrue();
+            expect(grid[0][1].startsWith('xf')).toBeTrue();
+            expect(grid[0][2].startsWith('xblank')).toBeTrue();
+            expect(grid[0][3].startsWith('xf')).toBeTrue();
+            expect(grid[0][0] === grid[0][1]).toBeTrue();
+
+            for (let row = 1; row < grid.length; row++) {
+                for (let col = 0; col < grid[0].length; col++) {
+                    expect(grid[row][col].startsWith('xblank')).toBeTrue();
+                }
+            }
+        });
+
+        it('grid layout compact UP hide ALL', () => {
+            expect(component.dataSource).toEqual([]);
+            expect(component.gridAreas).toBeUndefined();
+
+            component.computeLayoutData([
+                createMockDataGroup([
+                        createField(true, {x: 3, y: 0, rows: 1, cols: 1}),
+                        createField(true, {x: 0, y: 2, rows: 1, cols: 2})],
+                    undefined, DataGroupAlignment.START, DataGroupLayoutType.GRID, DataGroupCompact.UP, DataGroupHideEmptyRows.ALL)
+            ]);
+
+            expect(component.dataSource).toBeTruthy();
+            expect(Array.isArray(component.dataSource)).toBeTrue();
+            expect(component.dataSource.length).toEqual(3);
+
+            expect(component.gridAreas).toBeTruthy();
+            const grid = transformStringToGrid(component.gridAreas);
+
+            expect(grid.length).toEqual(1);
+            expect(grid[0].length).toEqual(4);
+
+            expect(grid[0][0].startsWith('xf')).toBeTrue();
+            expect(grid[0][1].startsWith('xf')).toBeTrue();
+            expect(grid[0][2].startsWith('xblank')).toBeTrue();
+            expect(grid[0][3].startsWith('xf')).toBeTrue();
+            expect(grid[0][0] === grid[0][1]).toBeTrue();
+        });
+
+        it('grid layout compact NONE hide NONE', () => {
+            expect(component.dataSource).toEqual([]);
+            expect(component.gridAreas).toBeUndefined();
+
+            component.computeLayoutData([
+                createMockDataGroup([
+                        createField(true, {x: 3, y: 0, rows: 1, cols: 1}),
+                        createField(true, {x: 0, y: 2, rows: 1, cols: 2})],
+                    undefined, DataGroupAlignment.START, DataGroupLayoutType.GRID, DataGroupCompact.NONE, DataGroupHideEmptyRows.NONE)
+            ]);
+
+            expect(component.dataSource).toBeTruthy();
+            expect(Array.isArray(component.dataSource)).toBeTrue();
+            expect(component.dataSource.length).toEqual(11);
+
+            expect(component.gridAreas).toBeTruthy();
+            const grid = transformStringToGrid(component.gridAreas);
+
+            expect(grid.length).toEqual(3);
+            expect(grid[0].length).toEqual(4);
+
+            expect(grid[0][0].startsWith('xblank')).toBeTrue();
+            expect(grid[0][1].startsWith('xblank')).toBeTrue();
+            expect(grid[0][2].startsWith('xblank')).toBeTrue();
+            expect(grid[0][3].startsWith('xf')).toBeTrue();
+
+            for (let col = 0; col < grid[0].length; col++) {
+                expect(grid[1][col].startsWith('xblank')).toBeTrue();
+            }
+
+            expect(grid[2][0].startsWith('xf')).toBeTrue();
+            expect(grid[2][1].startsWith('xf')).toBeTrue();
+            expect(grid[2][2].startsWith('xblank')).toBeTrue();
+            expect(grid[2][3].startsWith('xblank')).toBeTrue();
+            expect(grid[2][0] === grid[2][1]).toBeTrue();
+        });
+
+        it('grid layout compact NONE hide ALL', () => {
+            expect(component.dataSource).toEqual([]);
+            expect(component.gridAreas).toBeUndefined();
+
+            component.computeLayoutData([
+                createMockDataGroup([
+                        createField(true, {x: 3, y: 0, rows: 1, cols: 1}),
+                        createField(true, {x: 0, y: 2, rows: 1, cols: 2})],
+                    undefined, DataGroupAlignment.START, DataGroupLayoutType.GRID, DataGroupCompact.NONE, DataGroupHideEmptyRows.ALL)
+            ]);
+
+            expect(component.dataSource).toBeTruthy();
+            expect(Array.isArray(component.dataSource)).toBeTrue();
+            expect(component.dataSource.length).toEqual(7);
+
+            expect(component.gridAreas).toBeTruthy();
+            const grid = transformStringToGrid(component.gridAreas);
+
+            expect(grid.length).toEqual(2);
+            expect(grid[0].length).toEqual(4);
+
+            expect(grid[0][0].startsWith('xblank')).toBeTrue();
+            expect(grid[0][1].startsWith('xblank')).toBeTrue();
+            expect(grid[0][2].startsWith('xblank')).toBeTrue();
+            expect(grid[0][3].startsWith('xf')).toBeTrue();
+
+            expect(grid[1][0].startsWith('xf')).toBeTrue();
+            expect(grid[1][1].startsWith('xf')).toBeTrue();
+            expect(grid[1][2].startsWith('xblank')).toBeTrue();
+            expect(grid[1][3].startsWith('xblank')).toBeTrue();
+            expect(grid[1][0] === grid[1][1]).toBeTrue();
         });
     });